Redefine BoundingBox as a subclass & add category
Redefines BoundingBox to be composed of a Rectangle (identical to what was previously called BoundingBox), and a category.
